requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.2
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.3
six==1.15.0
sqlparse==0.4.2
whitenoise==5.3.0
